requirements.txt
django==4.0.*
environs==9.4.*
python-telegram-bot==13.10
